requirements.txt
# requirements.txt
# Instagram Scraper API Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Instagram Scraping
instaloader==4.11

# Async HTTP
aiohttp==3.9.1
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3

# ============================================
# .env.example
# Copy this to .env and configure

# API Configuration
API_KEYS=dev-key-12345,prod-key-67890

# Storage Configuration
DATA_DIR=./data

# Server Configuration
HOST=0.0.0.0
PORT=8000
WORKERS=4

# Logging
LOG_LEVEL=INFO

# Rate Limiting
RATE_LIMIT_REQUESTS=100
RATE_LIMIT_WINDOW=3600

# ============================================
# config.py
# Configuration management

import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()


class Config:
    """Application configuration"""
    
    # API Settings
    API_KEYS = set(os.getenv("API_KEYS", "dev-key-12345").split(","))
    HOST = os.getenv("HOST", "0.0.0.0")
    PORT = int(os.getenv("PORT", 8000))
    WORKERS = int(os.getenv("WORKERS", 4))
    
    # Storage Settings
    DATA_DIR = Path(os.getenv("DATA_DIR", "./data"))
    MAX_STORAGE_GB = int(os.getenv("MAX_STORAGE_GB", 10))
    
    # Scraping Settings
    MAX_URLS_PER_JOB = int(os.getenv("MAX_URLS_PER_JOB", 100))
    SCRAPE_DELAY_SECONDS = float(os.getenv("SCRAPE_DELAY_SECONDS", 2.0))
    MAX_COMMENTS_PER_POST = int(os.getenv("MAX_COMMENTS_PER_POST", 50))
    MAX_POSTS_PER_PROFILE = int(os.getenv("MAX_POSTS_PER_PROFILE", 12))
    
    # Rate Limiting
    RATE_LIMIT_REQUESTS = int(os.getenv("RATE_LIMIT_REQUESTS", 100))
    RATE_LIMIT_WINDOW = int(os.getenv("RATE_LIMIT_WINDOW", 3600))
    
    # Logging
    LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
    LOG_FILE = os.getenv("LOG_FILE", "scraper.log")
    
    @classmethod
    def validate(cls):
        """Validate configuration"""
        cls.DATA_DIR.mkdir(exist_ok=True)
        
        if not cls.API_KEYS:
            raise ValueError("At least one API key must be configured")


# ============================================
# Dockerfile
# Docker configuration for deployment

FROM python:3.11-slim

WORKDIR /app

# Install system dependencies
RUN apt-get update && apt-get install -y \
    gcc \
    && rm -rf /var/lib/apt/lists/*

# Copy requirements
COPY requirements.txt .

# Install Python dependencies
RUN pip install --no-cache-dir -r requirements.txt

# Copy application code
COPY . .

# Create data directory
RUN mkdir -p /app/data

# Expose port
EXPOSE 8000

# Health check
HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
    CMD curl -f http://localhost:8000/ || exit 1

# Run application
CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]

# ============================================
# docker-compose.yml
# Docker Compose for easy deployment

version: '3.8'

services:
  instagram-scraper:
    build: .
    ports:
      - "8000:8000"
    volumes:
      - ./data:/app/data
      - ./.env:/app/.env
    environment:
      - DATA_DIR=/app/data
    restart: unless-stopped
    healthcheck:
      test: ["CMD", "curl", "-f", "http://localhost:8000/"]
      interval: 30s
      timeout: 10s
      retries: 3
      start_period: 40s

# ============================================
# .gitignore
# Git ignore patterns

# Python
__pycache__/
*.py[cod]
*$py.class
*.so
.Python
env/
venv/
ENV/
build/
dist/
*.egg-info/

# Data
data/
*.json
*.csv
*.zip
*.jsonl

# Environment
.env
.env.local

# IDE
.vscode/
.idea/
*.swp
*.swo

# Logs
*.log
logs/

# OS
.DS_Store
Thumbs.db